Finish SymExStrategy
